Move the deprecatedTags stuff from OsmEntity to tags.js
It probably shouldnt be here either, but at least it is not
cluttering up the OsmEntity class.

For now, I set things up to work like the `osmAreaKeys` code.
This is something where the PresetSystem must first fetch the
data file at init time, then the data is set to a global export
so that other code can make use of it.