Add GeometryCollection and get Multipolygons and Relations working
Multipolygons and FeatureCollection relations were not working before
becuase the Geometry class could only handle simple types: point, line, polygon.

This commit introduces a GeometryCollection class to support mulipart
geometry, and adds code to extract the geometry parts from arbitrary
GeoJSON, so we can now support the multitypes and collections.
The Osm features now use this as the geometry store. (this.geom -> this.geoms)

This now means that GeoJSON is the intrnal format for handling geometry.
Before, we did a bunch of stuff to convert OSM Features into something
that the Geometry class could handle, and then considered `asGeoJSON()`
something that might get called later as needed.
Now, we just start out by getting into GeoJSON and then work from there.
We need the GeoJSON anyway, especially for the multipolygon and complex types.
And we will want this anyway to support non-OsmEntity feature types soon.