More changes to LocationSystem
We should move the initial resolution of the world location and
the blocked locations to `initAsync` because now that they create
`GeoJSON` data features, they depend on context being ready.

This change unfortunately breaks the `LocationSystem.test.js` by
introducing a strange `beforeEach` timeout issue with the returned promise,
but only when run under karma! It still works fine in a real Chrome browser.

I spent way too long trying to fix this, and I have no idea what the problem
is.  We even directly return a `Promise.resolve()` and mocha/karma
still experiences a timeout.