Do a little bit less work if the label debugging is off
This avoids creating and indexing debug boxes if the label debug is off.

It also restores the behavior of: When changing debug flags, dirty the scene.
(forcing the labels and debugging to be recomputed and rendered)

I also noticed that we still had several legacy context accessors for things
like `context.scene()` - I removed these and changed it so that any access
to the scene or layers really needs to go through the GraphicsSystem (gfx).