Formalize that all data types have an `updateGeometry()` method
Initially we really only needed this for OSM Ways and Relations,
which need a Graph in order to setup their geometry.  But we started
needing to do geometry replacement sometimes in other situations,
(streetlevel sequence lines getting updated as new markers load,
locations for markers that don't originally have them), so we
might as well accept that all data elements can rebuild their
geometry if they need to.