Attempting to avoid the mutable/immutable magic code in Graph
Graphs can be set up as 'mutable' but I thought this was only used
by the conflict-checking code when updating a temporary graph.

I found that actually the Graph.update() code creates
new Graphs as mutable, then runs the supplied funtions,
then switches the new graph back to being immutable.

The rotate and scale actions were also using Graph.update(),
but it didn't seem like there was any reason for this.

The problem I am really attacking is that I'd like to update geometry
(and tree) for Way/Relation entities after the graph is "consistent".
This means we need something like _updateCalculated for this step,
but it need to run after the Graph changes are applied.

I feel like there may be subtle bugs in the legacy code where
a Way or Relation gets a calculated transient (e.g. extent, geojson)
and that value is locked in (stored in the Graph), but then those
entities are still getting modified by either the `rebase` or `update`
calls and a wrong value is already locked in.