Move 'transient' from the Graph into the OsmEntities
This is a bit of an experiment.
"transients" are a cache where we memoize expensive calculations.

They previously lived in the Graph, however we churn through Graphs
pretty frequently, so they now live with the OsmEntities themselves.

The transients cache is shared between entities that copy from other entities.

For this to work, the Entity must be touched if there is a meaningful
change in the Graph that will cause the computed property to change.

The problem I am attacking right now is a performance regression in the
`checkAmbiguousCrossingTags` validator.  The way this validator works is
for each candidate way, it runs the "syncCrossingTags" action, and
checks the Difference from the start-end Graphs to see whether anything
has changed.

This has become problematic because, creating the new Graph and running
the action will now kick off a recalculation of the geometry, which is
expensive, particularly for multipolygon relations.  All this for
an action that probably doesn't change the way or multipolygon at all,
in most cases the difference will be empty.  So ideally, the Graph
should be checking the Difference and getting better at knowing
whether the changes have really affected the entities in a way that
require it to invalidate caches and geometry.
(Difference can kind of do this already, but Graph should use this)