WIP: Make Graph stateful and transactional
This is a big change to how Graphs work in Rapid.
This extends the work done in 2a572bc2 and 59b8558a,
and seems to be working pretty well.

Previously - We used Graphs in an immutable programming style, mostly.
Every `replace`, `remove`, `revert` would return a new Graph.

What really happened was that they would call `Graph.update`, which would
make the new Graph (as mutable), do the thing, update the caches, and then
set the Graph to be immutable.

Not only was this becoming a performance issue, but we had issues with
the caches needing to be updated in the correct order! Way and Relation
geometry can't be computed until `parentWays` and `parentRelations` are
settled.

So I'm moving from an immutable model to a transactional model, where code
can now apply a bunch of changes to the Graph, then call `commit()` once
it is ready.  `commit()` calculates the difference of what was done,
and updates the geometries on the affected entities.

This changes pretty much everywhere in the code and tests that
worked under the assumption that every replace/remove/revert returns
a new Graph.  The biggest change is in Actions, because their whole
purpose in the code is Graph manipulation.  I still want to keep
Actions "atomic", so we'll make sure for most actions, the last
thing it does is return a `graph.commit()`.

I'm still working through all the effects of this change, but at least
for now all the unit tests are passing.

Some tests needed to be rewritten becasue they now need to assume that
Graphs are mutable.  This isn't as scary as it sounds - if you want
to start changing stuff, just first do something like
  `graph = new Graph(predecessor)`
(which is what the EditSystem code does in practice anyway)
 or `graph = graph.commit()` would work too.