Move Graph, Difference, Tree into /models, Graph now accepts context.
These are more related to the data and should be with the data.
Graph now accepts a context, mostly so we can generate ids for each Graph.
Remove some dead code.

I'm working towards being able to chrun the Graphs less.
By having each graph link to the predecessor graph, it's essentially a memory leak.
We don't need to keep the link around after the geometries have been calculated.