Remove redundancy in the data models
This moves the `update` calls up to the base AbstractData class.
It's simpler and means we can remove a lot of redundant testing.

This also adds a destroy() method to Graph.
This also frees references to context in class destructors.
(I doubt this matters, but it might save the garbage collector a tiny bit of work)