Make the GeoJSON data work more like the other data element classes
Previously we just stuff the entire GeoJSON into the props,
and this was unusual.  It made this data a bit different
from working with OSM data and Markers.

This commit splits it so that the source GeoJSON data is now expected
to live in a `geojson` prop.  This approach allows us to store metadata
outside of that GeoJSON and return a fallback result for `asGeoJSON`
when source data isn't available.