WIP:  Add more methods/responsibility to SpatialSystem
This module can be responsible for keeping track of all spatial data,
and all coverage maps (tiles) of where that data has been loaded from.

I've started testing this out for some of the services (geoscribble, kartaview)
removing the caches and rbushes from the services and using the spatial system
to own these things instead.