Don't allow calls to replace/remove/revert on a base graph
The base graph should be for the unedited state.
You never want to call these methods on a base graph,
instead derive a working graph and perform the edits there.

This matches how we actually work with Graphs in Rapid,
however a lot of the tests were written in a way that base
graphs were getting modified, so this commit mostly rewrites
how the tests are constructed to allow the rules to be enforced.