Add epsilon for fuzzy comparison in rbush collision tests
This is because the floating point converstion between wgs84 and
world coordinates can incur a very slight drop in precision.
The difference wouldn't be noticible, but it could mean that:
orig = [lon, lat]
world = wgs84ToWorld(orig)
final = worldToWgs84(world)
orig !== final

This affects functions like `preventCoincidentLoc`, which might
return a loc that is slightly displaced from the original one.
This caused the function to miss some points that were effectively
 coincident but differed in this floating point error.