Remove area() function from OsmWay and OsmRelation
This function was only used by the `UiMeasurementCard` component
so I just moved that functionality over there.  There was also
some extra complexity to deal with winding order that we could
get rid of.  We can have the data classes generate normal GeoJSON,
then other parts of the code can perform computations if they need to.

Separately, the internal GeometryPart class now computes a planar
"area" in world coordinates, this is faster and useful enough for
knowing the relative size of things for rendering.