Stub out basic tests for all the rest of the core systems
This at least tests the constructor, initAsync, startAsync, resetAsync.

This commit also sorts out some more of the core system dependencies.
Many things still depend on other things, but we try to reduce the
connectivity between systems.  There are still a few surprises that
creep in as we are testing parts of the code.