Now systems can specify `requiredDependencies` and `optionalDependencies`
Before all dependencies were required, and a system would fail to
init if a dependency is not met.  Now we allow some dependencies to
be "optional".

Code must be written a way that allows a system to work without
an optional dependency, so we're using a lot of optional chaining.
For example: `urlhash?.getParam('thing') || ''`
Writing code like this means that the system should work fine even
in an environment that doesn't have a UrlHashSystem or mock for it.

Keeping the systems less interconnected helps a lot with testing.
Before every system would need a context containing almose every
other system (or mock it), and now we can test the systems with much
simpler mock environments.