Experiment: Replace EventEmitter3 with tseep
tseep claims to be a lot faster, but the main reason I want this
is because there are situations where I want to `prependListener`
and EventEmitter3 doesn't support this.

An odd side effect of tseep's performance optimizations is that it calls
the event callbacks with extra `undefined` arguments.  This doesn't affect
how the events work, but it does affect how our tests spy the events.
I needed to switch all the sinon spy calls from `calledOnceWithExactly`
to `calledOnceWith` because of these extra `undefined` args.