Remove the `loadedData`, `loadedNotes`, etc events
A long time ago in iD, the only thing we were interested in loading
was OSM data.  The OSM code would emit `loaded` when new data was
available, and this would trigger a map redraw.

In the years since we've added like a dozen more services to load data
and listening to `loadedData` events from all those modules became
cumbersome.  We've also switched to Pixi/WebGL which now has a ticker
running a game loop that does regular redraws.

So we stopped using the evented approach a while ago, and now any
service that has loaded some new data can request a redraw from
the graphics system.

This commit removes all the `loadedData` events that haven't been used
for anything in a while, becuase we should be testing how Rapid works
today.  The tests can spy on redraws just as easily.