More test work for services
1. Spy on redraws in some of the tests.
Followup from 52d7192c4, spy on redraws instead of `loadedData` events

2. Use xmldom for XML support.
I discovered that the loadedNotes tests were failing because our
OSM notes support still relies on XML, and our fetch_response code was
relying on the browser's built-in DOMParser to parse XML, which
isn't a thing built-in to Node when running tests.

I considered just switching this notes API code over to use the JSON API
(and we should still do that), but I also realized that we use XML for
supporting things like svg, kml, and gpx, and tmcw's togeojson libaray
recommends using `xmldom` instead of the browser's built-in DOMParser.
(see https://github.com/placemark/togeojson?tab=readme-ov-file#protips)
`xmldom` is a small dependency and it makes sense to use it where we need
XML support, so I added it the few places we need a `DOMParser`.

A few things with `xmldom` work a little bit differently,
so I needed to change the `parseNoteXML` code in a few places:
- Need to search for the root `osm` element instead of expecting it in
  `childNodes[0]` (`xmldom` gives us the `<?xml` processinginstruction
  in `childnodes[0]`, and a #text \n newline in `childNodes[1]`)
- `childNodes` is not iterable in v0.8.x of xmldom, but is in v0.9.x
  We can work around this by just `Array.from(childnodes)`.
  It is probably worth staying on v0.8.x for now because that's
  the version wanted by Pixi.js, and it would be dumb to bundle
  multiple versions of `xmldom` if we don't need to.

Again, we really should make sure most of the OSM API code is using
the JSON API and we can probably remove all the XML parsing code.
We still will need XML support anyway, like for parsing 'osc' files
or any of the other ones mentioned above, so `xmldom` will be useful.