Add GeoScribbleService tests, also remove Null Island tests
This adds tests for the GeoScribbleService.

I also decided to remove the testing for "does not load tiles around Null
Island", because this is an artifact of the Tiler which is tested elsewhere
in the rapid-sdk, and it doesn't meaningfully affect how the services work
or the test coverage of the service code.