More consistency in `immediateRedraw` and `deferredRedraw`
We had a lot of places where code would dispatch an event
and then it was MapSystem's job to catch that event and
trigger the redraw.  This shouldn't really be MapSystem's job.

Now the code is more consistent that any code can request redraws
by asking the GraphcsSystem for them.  It was like this before,
but now we are trying to do it everywhere like this.

GraphicsSystem is optional and might not exist when running
in a headless environment (no browsser / test).  So it's good
to add `gfx` as an optionaldependency and call the function
with null coalescing like `gfx?.immediateRedraw()`.

This means MapSystem doesn't need to mock out as much stuff.
I also got rid of the `immediateRedraw` and `deferredRedraw`
methods in MapSystem. These were legacy, only relayed the calls
to GraphicsSystem, and weren't called anywhere anymore.