Cleanup date handling code across Rapid
We had a lot of places with some duplicate code called something like
`_localeDateString`.  It was in many of the services, PhotoSystem, and
the ImagerySource class.  This code _either_ turned dates into an ISO
string like 'YYYY-MM-DD' _or_ a localized date like 'MMM YY, YYYY'.

I extracted this code into `utilDate` and `utilDateString` for
making things into dates and making strings like 'YYYY-MM-DD'.
I also added `displayShortDate` function to the LocalizationSystem
to handle the localized date situation.

This helps reduce the code duplication and improve testability.
(All this code is now tested now.)

It also makes it more straightforward to reduce dependence on the
LocalizationSystem.  For many situations (like testing) l10n can now
be an optional module that isn't installed or mocked.