Convert CLDR code to TypeScript, adjust linting rules
This converts the CLDR code to TypeScript.
This also adds linting to the ./scripts folder, which we didn't do before.
(And corrects the lint issues found there.)

This adjusts the linting configuration slightly to only lint
no-console and no-process-env in shipping code,
and allow these things in test and script code.