Add documentation, remove some localization code from preset/field
I'm trying to avoid having stuff like this in the fields and presets:
`t`, `tHtml`, `tAppend`, `hasTextForStringID`

It forces the l10n dependency on the preset system that makes testing more
difficult, and I want things as simple as possible to prove that things like
preset merging and inheritance works properly.

I'd like to get to the point where whatever caches we are using get
expired whenever a preset merge happens, or whenever a localization
switch happens.