Move Category/Preset/Field properties into `props` object
We were storing many of these in `orig` before, so this isn't a huge change.
But this makes them more like the data elements and other data items around
the Rapid code that work this way, also allows for more flexibility.
(If more props get added, we just have a place for them already).

This also fixes issues in the browser/ui/field tests.
These tests were just stuffing a Field into where a UiField should be,
and it just happened to work, but wasn't testing how the code works really.