Rename PresetSystem to SchemaSystem
This moves towards giving this class some more responsibility for
being the authority on tagging, being able to merge in multiple
tag-schema packs, and making that work flawlessly.

We already merge in id-tagging-schema, the rapid overrides, and the NSI,
so some of this capability already exists, but it is not tested.

The rename also settles a longtanding annoyance carried over from
the iD days where we use the word 'presets' to mean both the system,
but also the presets within the system.

This also changes the internal caches (_presets, _fields, _categories, etc)
to just be public properties.  It is kind of silly to make something
private, but then just put it behind a getter that lets anyone access it.
I'm trying to move away from code like this.

This also removes a few more places where we're using the Collection
class but we can just use an Array.  Collection's days are numbered!