Move the `/core/lib` and `data/lib` files into a common `/lib` folder
This makes things a little easier to manage.  All of these files will
work headlessly now, and I'm working on simplifying these helper classes
so that maybe there can be fewer of them.

This commit also goes through and simplifies the "barrel" `index.js` files a bit.
I've been trying to `export *` instead of doing named exports of everything.
This works ok because we know there are no longer name collisions in the Rapid codebase.
I've also been trying to avoid importing from index.js if it is just a single import.

pro: They are useful for making sure that everything gets exported into the
main bundle.

con: They can cause bundlers to be slower or unable to tree-shake properly.
This is less of an issue nowadays becuase bun is fast and I've made the code
more modular, free of side-effects and globals.  Also, most of this code is
actually needed, so tree-shakability is not really something I worry about much.