Move `search` from `Collection` directly into the `SchemaSystem`
Search was the last thing we needed the `Collection` for,
so this commit finally removes the `Collection` class.
It was just a wrapper around an `Array` that could do Preset searches,
and I've wanted to remove this extra complexity for a while.

The `search` tests are now moved into the `SchemaSystem` test suite.
These tests were already skipped and have been messed up for years.
I can now focus on getting them working again, and can think a bit
about whether we want to make any improvements to how preset search works.