Handle Category/Preset/Field string localization, caching
As the joke goes, the 2 hardest problems in computer science are
cache invalidation and naming things.  This commit hits both.

Now that we have MiniSearch fulltext index available to us...

1. Prelocalize all the Category/PresetField strings, so that they are
available for full-text search.  Previously, these things were
done "lazily" by the `name()` `terms()` etc functions, however
becuase these were getting called in the `search` function for
filtering and sorting, they were just getting called across the
entire array of Presets anyway - so not really lazy.

2. Break up the things that need to happen when locale changes (simple)
and when schema changes (destructive).

3. The strings and the searchIndex are now cached per-locale, so that
the user can switch languages while Rapid runs and things will just work.
The MiniSearch index will be rebuilt only if it needs to be.

4. Better handle the discrepency between `terms` and `aliases` , these
are both actually arrays, but the convention in Transifex is to represent
`terms` as comma-delimited and `aliases` as newline-delimited strings.

5. All of this locale code is written in a way so that the LocalizationSystem
is technically optional.  The SchemaSystem can do what it needs to do with
default properties, no translation, and fallback to `en-US` for everything.
As usual, trying to remove module interdependency to make things less
brittle and make testing easier.  No translations, no problem!