Switch from the old search code to MiniSearch
This removes a ton of old code that wasn't well tested
and replaces it with MiniSearch.  This gives us a better
search performance, some more control over how the
preset serach works, and makes it easier to support
localization of the preset strings.

Also removes the 'leven' dependency, since MiniSearch handles that.