Continue improving search
- In `SchemaSystem.search()`, combine search terms with 'AND'.
Default is 'OR' which means match _any_ of the terms extracted
from the query.
e.g.  'Home Depot' should match 'Home' AND 'Depot', not 'Home' OR 'Depot'

- Gathering tokens from the 'suggestion' presets was slow,
so this commit adds some code to gather the tokens more simply.

- Move unicode .normalize() into the script that fetches strings
from Transifex. This helps us avoid calling this at runtime
because we can assume all the strings are already normalized.