Improvement to Preset search, and expansion of test coverage
This improves the MiniSearch results slightly by performing the search
twice - once with exact matching and once with fuzzy matching.
We do this because:  we want exact matches to *always* beat fuzzy matches.
We don't want additional fuzzy matches to bump a result above an exact match.
For example:   query = "shop"
  "Shop" should be the top result.  "Shoe Shop" should never outrank it.

This also expands the test coverage to cover all the seraching
scenerios and handling of the recents and defaults.